requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
billiard==4.2.1
brotli==1.2.0
cachetools==6.2.0
celery==5.5.3
celery-types==0.23.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-pkg==0.15
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
cryptography==45.0.7
cssselect2==0.8.0
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==65.11.1
django-celery-beat==2.8.1
django-cors-headers==4.8.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-filter==25.1
django-jazzmin==3.0.1
django-redis==6.0.0
django-rest-framework==0.1.0
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
eskiz-sms==0.2.4
filelock==3.19.1
flower==2.0.1
fonttools==4.60.1
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
humanize==4.13.0
idna==3.10
inflection==0.5.1
jiter==0.10.0
kombu==5.5.4
oauthlib==3.3.1
openai==1.107.1
packaging==25.0
payme-pkg==3.0.29
pillow==11.3.0
platformdirs==4.4.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
PyJWT==2.10.1
pyparsing==3.2.5
pyphen==0.17.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
requests-mock==1.12.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.34.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
zopfli==0.4.0
